+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 <!-- here goes all the unreleased changes descriptions -->
+### Features
+- **Breaking Change**: emit in schema only types actually used by provided resolvers classes (#415)
+
 ### Fixes
 - refactor union types function syntax handling to prevent possible errors with circular refs